half precision teacher

+++ src/config.py
@@ -9,7 +9,7 @@ print(f"Using device: {DEVICE}")
 
 # Llama 3.2 1B
 MODEL = AutoModelForCausalLM.from_pretrained("meta-llama/Llama-3.2-1B-Instruct")
-MODEL.to(DEVICE)
+MODEL.half().to(DEVICE)
 
 TOKENIZER = AutoTokenizer.from_pretrained("meta-llama/Llama-3.2-1B-Instruct")
 if TOKENIZER.pad_token is None:
